move nav from components to layout
