--- web/src/components/MainLayoutNavbar/MainLayoutNavbar.stories.tsx
@@ -1,7 +1,0 @@
-import MainLayoutNavbar from './MainLayoutNavbar'
-
-export const generated = () => {
-  return <MainLayoutNavbar />
-}
-
-export default { title: 'Components/MainLayoutNavbar' }
