+++ web/src/layouts/MainLayout/Navigation/GithubStars/GithubStars.test.tsx
