+++ web/src/layouts/MainLayout/Navigation/GithubStars/GithubStars.tsx
